Move allocators into builder
We can split it out again as we implement more allocators.
